Split character mode and greek mode modules
