+++ src/compile/parseElements.ts
@@ -7,10 +7,8 @@ import {
 } from "../types.js";
 import type { PositionInfo } from "./buildPositionMap.js";
 import makeError from "../lib/makeError.js";
-import {
-  processCharacterMode,
-  processGreekMode,
-} from "./transliterateGreek.js";
+import processGreekMode from "./greekMode.js";
+import processCharacterMode from "./characterMode.js";
 
 export default (
   input: string,
